WaE 0xdeadbeaf 32 bits litteral casting to 64 bits pointer
Reviewed-on: https://gerrit.libreoffice.org/22600
Tested-by: Jenkins <ci@libreoffice.org>

+++ cppu/source/uno/prim.hxx
@@ -133,7 +133,7 @@ inline typelib_TypeDescriptionReference * _getVoidType()
 inline void CONSTRUCT_EMPTY_ANY(uno_Any * pAny) {
     pAny->pType = _getVoidType();
 #if OSL_DEBUG_LEVEL > 0
-    pAny->pData = reinterpret_cast<void *>(0xdeadbeef);
+    pAny->pData = reinterpret_cast<void *>((uintptr_t)0xdeadbeef);
 #else
     pAny->pData = pAny;
 #endif
